Added no ssr for the titlebar
