+++ apps/tauri/pages/_document.tsx
@@ -8,42 +8,13 @@ import NextDocument, {
 } from 'next/document'
 import { Children } from 'react'
 import { AppRegistry } from 'react-native'
-import { appWindow } from '@tauri-apps/api/window'
 
 import Tamagui from '../tamagui.config'
-import Image from 'next/image'
+import dynamic from 'next/dynamic'
 
-const TauriTitlebar = () => {
-  return (
-    <div data-tauri-drag-region className="titlebar">
-      <div className="titlebar-nav">
-        <div className="titlebar-button" id="titlebar-back" onClick={() => window.history.back()}>
-          <Image src="https://api.iconify.design/mdi:arrow-left.svg" alt="back" />
-        </div>
-        <div
-          className="titlebar-button"
-          id="titlebar-forward"
-          onClick={() => window.history.forward()}
-        >
-          <Image src="https://api.iconify.design/mdi:arrow-right.svg" alt="forward" />
-        </div>
-      </div>
-      <div className="titlebar-button" id="titlebar-minimize" onClick={() => appWindow.minimize()}>
-        <Image src="https://api.iconify.design/mdi:window-minimize.svg" alt="minimize" />
-      </div>
-      <div
-        className="titlebar-button"
-        id="titlebar-maximize"
-        onClick={() => appWindow.toggleMaximize()}
-      >
-        <Image src="https://api.iconify.design/mdi:window-maximize.svg" alt="maximize" />
-      </div>
-      <div className="titlebar-button" id="titlebar-close" onClick={() => appWindow.close()}>
-        <Image src="https://api.iconify.design/mdi:close.svg" alt="close" />
-      </div>
-    </div>
-  )
-}
+const TauriTitlebar = dynamic(() => import('components/TauriTitlebar'), {
+  ssr: false,
+})
 
 export default class Document extends NextDocument {
   static async getInitialProps(ctx: DocumentContext): Promise<DocumentInitialProps> {
